refactor: load theme script externally
