+++ public/theme-init.js
@@ -1,0 +1,10 @@
+(function() {
+  try {
+    var storageKey = 'theme';
+    var stored = localStorage.getItem(storageKey);
+    var prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+    if (stored === 'dark' || (!stored && prefersDark)) {
+      document.documentElement.classList.add('dark');
+    }
+  } catch (e) {}
+})();
